requirements.txt
fastapi==0.112.0
websockets==13.0.1
uvicorn
pydantic==2.9.0
pydantic-settings==2.8.0
python-dotenv==1.0.1
setuptools==71.0.2
azure-common==1.1.28
azure-core==1.30.2
azure-identity==1.17.1
azure-mgmt-core==1.4.0
azure-mgmt-monitor==6.0.2
azure-monitor-opentelemetry==1.6.1
azure-search-documents==11.5.1
azure-storage-blob==12.22.0
azure-storage-queue==12.11.0
SQLAlchemy==2.0.32
pyodbc
alembic==1.13.2
tenacity==8.5.0
langchain==0.2.13
langchain-community==0.2.12
langchain-core==0.2.30
langchain-openai==0.1.21
langchain-text-splitters==0.2.2
pillow==10.4.0
azure-functions
tqdm==4.66.5
azure-ai-formrecognizer==3.3.0
azure-ai-documentintelligence==1.0.0b3
pymupdf==1.24.10
python-magic==0.4.27
python-pptx==1.0.2
pandas==2.2.3
openpyxl==3.1.5
python-slugify==8.0.4
unstructured==0.16.23
networkx==3.4.2
structlog==24.4.0
pdfminer.six==20231228
pdfplumber==0.11.5
comtypes==1.4.10
azure-functions
psycopg2==2.9.10
ragas==0.2.14